intervalcourses/intervalcourses/requirements.txt
Django==4.2.9
django-admin-interface==0.29.4
django-ckeditor-5==0.2.17
django-colorfield==0.12.0
django-countries==7.6.1
django-extensions==3.2.3
django-honeypot==1.2.1
django-import-export==4.3.7
django-multiselectfield==0.1.13
django-ratelimit==4.1.0
django-taggit==6.1.0
easy-thumbnails==2.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-slugify==8.0.4
razorpay==1.4.2
requests==2.32.3
sentry-sdk==2.23.1